# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Car Repair Register API",
    "description": "Evidence of customers and vehicles for small workshops.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/api/car-repair-register",
      "description": "Default Server URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/repair-log/{id}": {
      "get": {
        "tags": [
          "repairLog"
        ],
        "description": "Get repair log by id.",
        "operationId": "getRepairLogById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "ok",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RepairLog"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "repairLog"
        ],
        "description": "Update existing repair log.",
        "operationId": "updateRepairLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairLogUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "ok",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RepairLog"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "repairLog"
        ],
        "description": "Delete repair log with all its records.",
        "operationId": "deleteRepairLogById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/repair-log": {
      "get": {
        "tags": [
          "repairLog"
        ],
        "description": "Find repair logs by query string, vehicleId and pageable.",
        "operationId": "findRepairLogs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "name": "size",
            "in": "query",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "name": "sort",
            "in": "query",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "ok",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModel"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repairLog"
        ],
        "description": "Create new repair log.",
        "operationId": "createRepairLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "201": {
            "description": "created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RepairLog"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "required": [
          "fieldError",
          "httpStatus",
          "message"
        ],
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "string",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 EARLY_HINTS",
              "103 CHECKPOINT",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "302 MOVED_TEMPORARILY",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "305 USE_PROXY",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 PAYLOAD_TOO_LARGE",
              "413 REQUEST_ENTITY_TOO_LARGE",
              "414 URI_TOO_LONG",
              "414 REQUEST_URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "419 INSUFFICIENT_SPACE_ON_RESOURCE",
              "420 METHOD_FAILURE",
              "421 DESTINATION_LOCKED",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "message": {
            "type": "string"
          },
          "fieldError": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "required": [
          "errorMessage",
          "fieldName"
        ],
        "type": "object",
        "properties": {
          "fieldName": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "nullable": true
      },
      "Customer": {
        "required": [
          "created",
          "email",
          "id",
          "mobile",
          "modified",
          "name",
          "surname"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Vehicle": {
        "required": [
          "batteryCapacity",
          "brand",
          "created",
          "customer",
          "engineCode",
          "enginePower",
          "engineVolume",
          "fuelType",
          "id",
          "model",
          "modified",
          "registrationPlate",
          "vin",
          "yearOfManufacture"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "registrationPlate": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "engineCode": {
            "type": "string",
            "nullable": true
          },
          "fuelType": {
            "type": "string",
            "nullable": true
          },
          "enginePower": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "engineVolume": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "batteryCapacity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "yearOfManufacture": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "RepairLogUpdate": {
        "required": [
          "content",
          "repairDate",
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string"
          },
          "vehicleId": {
            "type": "integer",
            "format": "int64"
          },
          "repairDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Attachment": {
        "required": [
          "created",
          "id",
          "mimeType",
          "modified",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "RepairLog": {
        "required": [
          "attachments",
          "content",
          "created",
          "id",
          "modified",
          "repairDate",
          "vehicle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "vehicle": {
            "$ref": "#/components/schemas/Vehicle"
          },
          "repairDate": {
            "type": "string",
            "format": "date"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "RepairLogCreate": {
        "required": [
          "content",
          "repairDate",
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string"
          },
          "vehicleId": {
            "type": "integer",
            "format": "int64"
          },
          "repairDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModel": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}